requirements.txt
streamlit
transformers
pdfminer.six
torch
nltk
PyMuPDF
scikit-learn
numpy
python-docx